+++ playground/hoa0217/java-springboot-order-perf-multi-module/display/src/main/java/com/example/common/CustomCacheManager.java
@@ -1,4 +1,4 @@
-package com.example.config;
+package com.example.common;
 
 public interface CustomCacheManager {
