+++ playground/hoa0217/java-springboot-order-perf-multi-module/display/src/main/java/com/example/common/LocalCacheManager.java
@@ -1,4 +1,4 @@
-package com.example.config;
+package com.example.common;
 
 import java.util.Objects;
 import lombok.RequiredArgsConstructor;
